Restructure F6 docs and tests
