+++ features/F6/acceptance_tests/acceptance.py
@@ -1,0 +1,13 @@
+from .s1 import f6s1
+from .s2 import f6s2
+from .s3 import f6s3
+from .s4 import f6s4
+from .s5 import f6s5
+
+__all__ = [
+    "f6s1",
+    "f6s2",
+    "f6s3",
+    "f6s4",
+    "f6s5",
+]
