+++ features/F6/specification.md
@@ -84,10 +84,10 @@ services:
 
 | #     | Scenario & pre‑conditions                                      | Steps (user actions → expected behaviour)                                                                                                                                                                 |
 | ----- | -------------------------------------------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| **1** | **Add a file**<br>Stack running and WebDAV endpoint reachable. | 1 `curl -T ./a.txt http://localhost:8000/dav/a.txt` → `input/metadata/by-id/<hash>/document.json` exists **and** searching for that `id` returns a document without waiting for a cron tick. ([test](../features/F6/test/acceptance.py#L93-L112)) |
-| **2** | **Move a file (JSON)**                                         | 1 After Scenario 1 POST<br>`{"move":[{"src":"a.txt","dest":"b.txt"}]}` to `/fileops` → Document’s `paths` contain `b.txt` and no longer include `a.txt`; search results show the new path within seconds. ([test](../features/F6/test/acceptance.py#L115-L144)) |
-| **3** | **Delete a file**                                              | 1 POST `{"delete":["b.txt"]}` to `/fileops` → Document disappears from both metadata store and search index. ([test](../features/F6/test/acceptance.py#L147-L175)) |
-| **4** | **Batch operations**                                           | 1 Upload two files via `/dav` then POST a single body<br>`{"move":[{"src":"c.txt","dest":"e.txt"}], "delete": ["d.txt"]}` → All changes reflected correctly after the response. ([test](../features/F6/test/acceptance.py#L178-L221)) |
-| **5** | **Rename via WebDAV**                                          | 1 `curl -X MOVE -H 'Destination: http://localhost:8000/dav/f.txt' http://localhost:8000/dav/e.txt` → Same results as Scenario 2, proving that WebDAV MOVE is equivalent to the JSON `move`. ([test](../features/F6/test/acceptance.py#L224-L251)) |
+| **1** | **Add a file**<br>Stack running and WebDAV endpoint reachable. | 1 `curl -T ./a.txt http://localhost:8000/dav/a.txt` → `input/metadata/by-id/<hash>/document.json` exists **and** searching for that `id` returns a document without waiting for a cron tick. ([test](acceptance_tests/s1.py)) |
+| **2** | **Move a file (JSON)**                                         | 1 After Scenario 1 POST<br>`{"move":[{"src":"a.txt","dest":"b.txt"}]}` to `/fileops` → Document’s `paths` contain `b.txt` and no longer include `a.txt`; search results show the new path within seconds. ([test](acceptance_tests/s2.py)) |
+| **3** | **Delete a file**                                              | 1 POST `{"delete":["b.txt"]}` to `/fileops` → Document disappears from both metadata store and search index. ([test](acceptance_tests/s3.py)) |
+| **4** | **Batch operations**                                           | 1 Upload two files via `/dav` then POST a single body<br>`{"move":[{"src":"c.txt","dest":"e.txt"}], "delete": ["d.txt"]}` → All changes reflected correctly after the response. ([test](acceptance_tests/s4.py)) |
+| **5** | **Rename via WebDAV**                                          | 1 `curl -X MOVE -H 'Destination: http://localhost:8000/dav/f.txt' http://localhost:8000/dav/e.txt` → Same results as Scenario 2, proving that WebDAV MOVE is equivalent to the JSON `move`. ([test](acceptance_tests/s5.py)) |
 
 All scenarios **must pass unchanged** on Linux, macOS, and Windows (WSL).
